requirements.txt
tensorflow==2.15.0
numpy>=1.23.5,<2.0.0
PyYAML==6.0.3
tensorflow-datasets==4.9.9
matplotlib
pytest>=7.0.0
torch
facenet-pytorch
importlib-resources